Do not include quotes in PATH
There can be parsing errors, when there are quotes in the PATH environment variable. Unless the cuda_path contains spaces the quotes shouldn't be needed and cuda_path should never contain spaces, as that would be rather dangerous. (And in practise if you install cuda from the debian repository it will never contain spaces.)